refactor: extract shared pagination types into agentd-common crate
Move duplicated PaginatedResponse<T>, clamp_limit(), DEFAULT_PAGE_LIMIT,
and MAX_PAGE_LIMIT from notify and orchestrator into the agentd-common
crate's types module.

Both notify/src/types.rs and orchestrator/src/types.rs now re-export
from agentd_common::types, preserving all existing imports and API
surface for downstream consumers (cli, tests).

Changes:
- crates/common/src/types.rs: populated with PaginatedResponse<T>,
  pagination constants, clamp_limit(), doc-tests, and 5 unit tests
- crates/notify/Cargo.toml: add agentd-common dependency
- crates/notify/src/types.rs: replace inline definitions with re-exports
- crates/orchestrator/Cargo.toml: add agentd-common dependency
- crates/orchestrator/src/types.rs: replace inline definitions with
  re-exports

No functional changes — the types, serialization format, and API are
identical. All existing imports in api.rs, client.rs, scheduler/api.rs,
and CLI continue to work unchanged.

Closes #45
