+++ crates/notify/Cargo.toml
@@ -38,6 +38,7 @@ sqlx = { version = "0.8", features = [
 uuid = { version = "1.11", features = ["v4", "serde"] }
 chrono = { version = "0.4", features = ["serde"] }
 directories = "6.0.0"
+agentd-common = { path = "../common" }
 metrics = { workspace = true }
 metrics-exporter-prometheus = { workspace = true }
