+++ crates/notify/src/types.rs
@@ -602,28 +602,8 @@ pub struct CountResponse {
 ///   "offset": 0
 /// }
 /// ```
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PaginatedResponse<T> {
-    /// The items in this page.
-    pub items: Vec<T>,
-    /// Total number of items matching the query (across all pages).
-    pub total: usize,
-    /// Maximum items per page.
-    pub limit: usize,
-    /// Offset from the start.
-    pub offset: usize,
-}
-
-/// Default page size when `limit` is not specified.
-pub const DEFAULT_PAGE_LIMIT: usize = 50;
-
-/// Maximum allowed page size.
-pub const MAX_PAGE_LIMIT: usize = 200;
-
-/// Clamp a requested limit to valid bounds.
-pub fn clamp_limit(limit: Option<usize>) -> usize {
-    limit.unwrap_or(DEFAULT_PAGE_LIMIT).clamp(1, MAX_PAGE_LIMIT)
-}
+// Re-export pagination types from agentd-common.
+pub use agentd_common::types::{clamp_limit, PaginatedResponse, DEFAULT_PAGE_LIMIT, MAX_PAGE_LIMIT};
 
 #[cfg(test)]
 mod tests {
